// src/app/interface/ferias.interface.ts
export interface FeriasInterface {
    cpf: string;
    tipoSalario: string;
    salarioBase: number;
    jornada: number;
    dependentesIR: number;
    percenPensao: number;
    ehSalarioPorHora: boolean;
    ehFeriasEmDobro: boolean;
    ehCom13Salario: boolean;
    diasFerias: number;
    diasAbono: number;
    mediaSalarioVariavel: number;
    dataFerias: string;
}